Make new user token clipboard copyable
