+++ http/html/static/templates/content/token_created.tmpl
@@ -1,0 +1,1 @@
+Created token: <span class="data clipboard-copyable">{{ . }}</span>
